+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 * Documentation: Added in [crate::stdlib] a rough documentation of the available variables and nodes.
 * Change: Refactored the use of auxilary variables. Sample rate variables are not
 prefixed with an `$` (`$srate` and `$israte`).
+* Change: Persistent variables will now also be resetted to 0.0 after a reset or first initialization.
 * Feature: Added reading/writing into a shared buffer of atomic floats.
 See also `DSPNodeContext::atom` and `CodeEngine::atom`. Also added an example
 that shows how to control a DSPFunction with atoms.
 * Feature: Added constants for PI, E, TAU and many others.
+* Feature: Added `$reset` which is true directly after an (explicit) reset and the first initialization.
 
 0.5.3 (2022-08-05)
 ==================